backend/pyproject.toml
[tool.poetry]
name = "django-rest-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.3"
djangorestframework = "^3.15.2"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
djoser = "^2.3.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
